Move lastfm code into own folder
