+++ internal/clients/lastfm/interfaces.go
@@ -1,4 +1,4 @@
-package clients
+package lastfm
 
 type LastFMResponse interface {
 	Append(l LastFMResponse) error
